refactor: split TestCaseNode computation from data
To allow TestCaseNode to be reused by other computation, split the computation out and leave a Pojo.
