+++ ksql-functional-tests/src/main/java/io/confluent/ksql/test/tools/TestCaseBuilder.java
@@ -1,0 +1,349 @@
+/*
+ * Copyright 2019 Confluent Inc.
+ *
+ * Licensed under the Confluent Community License (the "License"); you may not use
+ * this file except in compliance with the License.  You may obtain a copy of the
+ * License at
+ *
+ * http://www.confluent.io/confluent-community-license
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package io.confluent.ksql.test.tools;
+
+import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableMap;
+import com.google.common.collect.Iterables;
+import com.google.common.collect.Streams;
+import io.confluent.connect.avro.AvroData;
+import io.confluent.ksql.function.FunctionRegistry;
+import io.confluent.ksql.function.TestFunctionRegistry;
+import io.confluent.ksql.metastore.MetaStoreImpl;
+import io.confluent.ksql.metastore.model.KsqlTopic;
+import io.confluent.ksql.model.WindowType;
+import io.confluent.ksql.parser.DefaultKsqlParser;
+import io.confluent.ksql.parser.KsqlParser;
+import io.confluent.ksql.parser.KsqlParser.ParsedStatement;
+import io.confluent.ksql.parser.KsqlParser.PreparedStatement;
+import io.confluent.ksql.parser.SqlBaseParser;
+import io.confluent.ksql.parser.tree.CreateSource;
+import io.confluent.ksql.schema.ksql.LogicalSchema;
+import io.confluent.ksql.schema.ksql.SchemaConverters;
+import io.confluent.ksql.serde.Format;
+import io.confluent.ksql.serde.FormatInfo;
+import io.confluent.ksql.serde.KeyFormat;
+import io.confluent.ksql.serde.ValueFormat;
+import io.confluent.ksql.serde.WindowInfo;
+import io.confluent.ksql.test.model.PostConditionsNode;
+import io.confluent.ksql.test.model.TestCaseNode;
+import io.confluent.ksql.test.model.WindowData;
+import io.confluent.ksql.test.model.WindowData.Type;
+import io.confluent.ksql.test.serde.SerdeSupplier;
+import io.confluent.ksql.test.tools.conditions.PostConditions;
+import io.confluent.ksql.test.tools.exceptions.InvalidFieldException;
+import io.confluent.ksql.test.tools.exceptions.KsqlExpectedException;
+import io.confluent.ksql.test.utils.SerdeUtil;
+import io.confluent.ksql.topic.TopicFactory;
+import io.confluent.ksql.util.DecimalUtil;
+import io.confluent.ksql.util.KsqlConstants;
+import java.nio.file.Path;
+import java.time.Duration;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
+import java.util.Optional;
+import java.util.UUID;
+import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.function.Supplier;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import org.apache.kafka.connect.data.Field;
+import org.apache.kafka.connect.data.Schema;
+import org.apache.kafka.connect.data.SchemaBuilder;
+
+/**
+ * Builds {@link TestCase}s from {@link TestCaseNode}.
+ */
+@SuppressWarnings("UnstableApiUsage")
+public final class TestCaseBuilder {
+
+  private static final String FORMAT_REPLACE_ERROR =
+      "To use {FORMAT} in your statements please set the 'format' test case element";
+
+  private final FunctionRegistry functionRegistry;
+
+  public TestCaseBuilder() {
+    this(TestFunctionRegistry.INSTANCE.get());
+  }
+
+  public TestCaseBuilder(final FunctionRegistry functionRegistry) {
+    this.functionRegistry = Objects.requireNonNull(functionRegistry, "functionRegistry");
+  }
+
+  public List<TestCase> buildTests(final TestCaseNode test, final Path testPath) {
+    if (!test.isEnabled()) {
+      return ImmutableList.of();
+    }
+
+    try {
+      final Stream<Optional<String>> formats = test.formats().isEmpty()
+          ? Stream.of(Optional.empty())
+          : test.formats().stream().map(Optional::of);
+
+      return formats
+          .map(format -> createTest(test, format, testPath))
+          .collect(Collectors.toList());
+    } catch (final Exception e) {
+      throw new AssertionError("Invalid test '" + test.name() + "': " + e.getMessage(), e);
+    }
+  }
+
+  private TestCase createTest(
+      final TestCaseNode test,
+      final Optional<String> explicitFormat,
+      final Path testPath
+  ) {
+    final String testName = buildTestName(testPath, test.name(), explicitFormat);
+
+    try {
+      final List<String> statements = buildStatements(test.statements(), explicitFormat);
+
+      final Optional<KsqlExpectedException> ee = test.expectedException()
+          .map(een -> een.build(Iterables.getLast(statements)));
+
+      final Map<String, Topic> topics = getTestCaseTopics(
+          test,
+          statements,
+          explicitFormat,
+          ee.isPresent()
+      );
+
+      final List<Record> inputRecords = test.inputs().stream()
+          .map(node -> node.build(topics))
+          .collect(Collectors.toList());
+
+      final List<Record> outputRecords = test.outputs().stream()
+          .map(node -> node.build(topics))
+          .collect(Collectors.toList());
+
+      final PostConditions post = test.postConditions()
+          .map(PostConditionsNode::build)
+          .orElse(PostConditions.NONE);
+
+      return new TestCase(
+          testPath,
+          testName,
+          Optional.empty(),
+          test.properties(),
+          topics.values(),
+          inputRecords,
+          outputRecords,
+          statements,
+          ee.orElseGet(KsqlExpectedException::none),
+          post
+      );
+    } catch (final Exception e) {
+      throw new AssertionError(testName + ": Invalid test. " + e.getMessage(), e);
+    }
+  }
+
+  private static List<String> buildStatements(
+      final List<String> statements,
+      final Optional<String> explicitFormat
+  ) {
+    final String format = explicitFormat.orElse(FORMAT_REPLACE_ERROR);
+
+    return statements.stream()
+        .map(stmt -> stmt.replace("{FORMAT}", format))
+        .collect(Collectors.toList());
+  }
+
+  private Map<String, Topic> getTestCaseTopics(
+      final TestCaseNode test,
+      final List<String> statements,
+      final Optional<String> defaultFormat,
+      final boolean expectsException
+  ) {
+    final Map<String, Topic> allTopics = new HashMap<>();
+
+    // Add all topics from topic nodes to the map:
+    test.topics().stream()
+        .map(node -> node.build(defaultFormat))
+        .forEach(topic -> allTopics.put(topic.getName(), topic));
+
+    // Infer topics if not added already:
+    statements.stream()
+        .map(this::createTopicFromStatement)
+        .filter(Objects::nonNull)
+        .forEach(topic -> allTopics.putIfAbsent(topic.getName(), topic));
+
+    if (allTopics.isEmpty()) {
+      if (expectsException) {
+        return ImmutableMap.of();
+      }
+      throw new InvalidFieldException("statements/topics", "The test does not define any topics");
+    }
+
+    final SerdeSupplier<?> defaultValueSerdeSupplier =
+        allTopics.values().iterator().next().getValueSerdeSupplier();
+
+    // Get topics from inputs and outputs fields:
+    Streams.concat(test.inputs().stream(), test.outputs().stream())
+        .map(recordNode -> new Topic(
+            recordNode.topicName(),
+            Optional.empty(),
+            getKeySedeSupplier(recordNode.getWindow()),
+            defaultValueSerdeSupplier,
+            4,
+            1,
+            Optional.empty()))
+        .forEach(topic -> allTopics.putIfAbsent(topic.getName(), topic));
+
+    return allTopics;
+  }
+
+  private static SerdeSupplier<?> getKeySedeSupplier(final Optional<WindowData> windowDataInfo) {
+    if (windowDataInfo.isPresent()) {
+      final WindowData windowData = windowDataInfo.get();
+      final WindowType windowType = WindowType.of((windowData.type == Type.SESSION)
+          ? WindowType.SESSION.name()
+          : WindowType.TUMBLING.name());
+      final KeyFormat windowKeyFormat = KeyFormat.windowed(
+          Format.KAFKA,
+          Optional.empty(),
+          WindowInfo.of(
+              windowType,
+              windowType == WindowType.SESSION
+                  ? Optional.empty() : Optional.of(Duration.ofMillis(windowData.size())))
+      );
+      return SerdeUtil.getKeySerdeSupplier(windowKeyFormat, () -> LogicalSchema.builder().build());
+    }
+    return SerdeUtil.getKeySerdeSupplier(
+        KeyFormat.nonWindowed(FormatInfo.of(Format.KAFKA)),
+        () -> LogicalSchema.builder().build());
+  }
+
+  private Topic createTopicFromStatement(final String sql) {
+    final KsqlParser parser = new DefaultKsqlParser();
+    final MetaStoreImpl metaStore = new MetaStoreImpl(functionRegistry);
+
+    final Predicate<ParsedStatement> onlyCSandCT = stmt ->
+        stmt.getStatement().statement() instanceof SqlBaseParser.CreateStreamContext
+            || stmt.getStatement().statement() instanceof SqlBaseParser.CreateTableContext;
+
+    final Function<PreparedStatement<?>, Topic> extractTopic = (PreparedStatement<?> stmt) -> {
+      final CreateSource statement = (CreateSource) stmt.getStatement();
+
+      final KsqlTopic ksqlTopic = TopicFactory.create(statement.getProperties());
+
+      final KeyFormat keyFormat = ksqlTopic.getKeyFormat();
+
+      final Supplier<LogicalSchema> logicalSchemaSupplier =
+          statement.getElements()::toLogicalSchema;
+
+      final SerdeSupplier<?> keySerdeSupplier =
+          SerdeUtil.getKeySerdeSupplier(keyFormat, logicalSchemaSupplier);
+
+      final ValueFormat valueFormat = ksqlTopic.getValueFormat();
+      final Optional<org.apache.avro.Schema> avroSchema;
+      if (valueFormat.getFormat() == Format.AVRO) {
+        // add avro schema
+        final SchemaBuilder schemaBuilder = SchemaBuilder.struct();
+        statement.getElements().forEach(e -> schemaBuilder.field(
+            e.getName(),
+            SchemaConverters.sqlToConnectConverter().toConnectSchema(e.getType().getSqlType()))
+        );
+        avroSchema = Optional.of(new AvroData(1)
+            .fromConnectSchema(addNames(schemaBuilder.build())));
+      } else {
+        avroSchema = Optional.empty();
+      }
+
+      final SerdeSupplier<?> valueSerdeSupplier = SerdeUtil.getSerdeSupplier(
+          valueFormat.getFormat(),
+          logicalSchemaSupplier
+      );
+
+      return new Topic(
+          ksqlTopic.getKafkaTopicName(),
+          avroSchema,
+          keySerdeSupplier,
+          valueSerdeSupplier,
+          KsqlConstants.legacyDefaultSinkPartitionCount,
+          KsqlConstants.legacyDefaultSinkReplicaCount,
+          Optional.empty());
+    };
+
+    try {
+      final List<ParsedStatement> parsed = parser.parse(sql);
+      if (parsed.size() > 1) {
+        throw new IllegalArgumentException("SQL contains more than one statement: " + sql);
+      }
+
+      final List<Topic> topics = parsed.stream()
+          .filter(onlyCSandCT)
+          .map(stmt -> parser.prepare(stmt, metaStore))
+          .map(extractTopic)
+          .collect(Collectors.toList());
+
+      return topics.isEmpty() ? null : topics.get(0);
+    } catch (final Exception e) {
+      // Statement won't parse: this will be detected/handled later.
+      System.out.println("Error parsing statement (which may be expected): " + sql);
+      e.printStackTrace(System.out);
+      return null;
+    }
+  }
+
+  private static Schema addNames(final Schema schema) {
+    final SchemaBuilder builder;
+    switch (schema.type()) {
+      case BYTES:
+        DecimalUtil.requireDecimal(schema);
+        builder = DecimalUtil.builder(schema);
+        break;
+      case ARRAY:
+        builder = SchemaBuilder.array(addNames(schema.valueSchema()));
+        break;
+      case MAP:
+        builder = SchemaBuilder.map(
+            Schema.STRING_SCHEMA,
+            addNames(schema.valueSchema())
+        );
+        break;
+      case STRUCT:
+        builder = SchemaBuilder.struct();
+        builder.name("TestSchema" + UUID.randomUUID().toString().replace("-", ""));
+        for (final Field field : schema.fields()) {
+          builder.field(field.name(), addNames(field.schema()));
+        }
+        break;
+      default:
+        builder = SchemaBuilder.type(schema.type());
+    }
+    if (schema.isOptional()) {
+      builder.optional();
+    }
+    return builder.build();
+  }
+
+  private static String buildTestName(
+      final Path testPath,
+      final String testName,
+      final Optional<String> postfix
+  ) {
+    final String fileName = com.google.common.io.Files.getNameWithoutExtension(testPath.toString());
+
+    final String pf = postfix
+        .map(f -> " - " + f)
+        .orElse("");
+
+    return fileName + " - " + testName + pf;
+  }
+}
