+++ ksql-functional-tests/src/test/java/io/confluent/ksql/test/QueryTranslationTest.java
@@ -20,12 +20,12 @@ import static java.util.Objects.requireNonNull;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.collect.ImmutableList;
-import io.confluent.ksql.function.TestFunctionRegistry;
 import io.confluent.ksql.test.loader.ExpectedTopologiesTestLoader;
 import io.confluent.ksql.test.loader.JsonTestLoader;
 import io.confluent.ksql.test.loader.TestFile;
 import io.confluent.ksql.test.model.TestCaseNode;
 import io.confluent.ksql.test.tools.TestCase;
+import io.confluent.ksql.test.tools.TestCaseBuilder;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -85,6 +85,7 @@ public class QueryTranslationTest {
   static class QttTestFile implements TestFile<TestCase> {
 
     private final List<TestCaseNode> tests;
+    private final TestCaseBuilder builder = new TestCaseBuilder();
 
     QttTestFile(@JsonProperty("tests") final List<TestCaseNode> tests) {
       this.tests = ImmutableList.copyOf(requireNonNull(tests, "tests collection missing"));
@@ -98,7 +99,7 @@ public class QueryTranslationTest {
     public Stream<TestCase> buildTests(final Path testPath) {
       return tests
           .stream()
-          .flatMap(node -> node.buildTests(testPath, TestFunctionRegistry.INSTANCE.get()).stream());
+          .flatMap(node -> builder.buildTests(node, testPath).stream());
     }
   }
 }
