+++ ksql-functional-tests/src/test/java/io/confluent/ksql/test/tools/TestExecutorUtilTest.java
@@ -21,7 +21,6 @@ import static org.hamcrest.MatcherAssert.assertThat;
 
 import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientException;
 import io.confluent.ksql.engine.KsqlEngine;
-import io.confluent.ksql.function.TestFunctionRegistry;
 import io.confluent.ksql.json.JsonMapper;
 import io.confluent.ksql.services.ServiceContext;
 import io.confluent.ksql.test.model.QttTestFile;
@@ -53,9 +52,8 @@ public class TestExecutorUtilTest {
     final QttTestFile qttTestFile = JsonMapper.INSTANCE.mapper
         .readValue(new File("src/test/resources/testing_tool_tests.json"), QttTestFile.class);
     final TestCaseNode testCaseNode = qttTestFile.tests.get(0);
-    testCase = testCaseNode.buildTests(
-        new File("src/test/resources/testing_tool_tests.json").toPath(),
-        TestFunctionRegistry.INSTANCE.get()
+    testCase = new TestCaseBuilder().buildTests(testCaseNode,
+        new File("src/test/resources/testing_tool_tests.json").toPath()
     ).get(0);
 
     serviceContext = TestExecutor.getServiceContext();
